pyproject.toml
[tool.poetry]
name = "goit-python-web-hw-12"
version = "0.1.0"
description = ""
authors = ["Serghei <siracencoserghei@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0.post1"}
sqlalchemy = "^2.0.23"
psycopg2 = "^2.9.9"
pydantic = {extras = ["email"], version = "^2.4.2"}
alembic = "^1.12.1"
jinja2 = "^3.1.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
